requirements.txt
cloudpickle==3.0.0
dotmap==1.3.30
Farama-Notifications==0.0.4
filelock==3.14.0
fsspec==2024.3.1
gymnasium==0.29.1
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
pillow==10.3.0
pygame==2.5.2
PyYAML==6.0.1
sympy==1.12
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
triton==2.3.0
typing_extensions==4.11.0
